FakturaserieService skal ikke ha tilgang til fakturaGenerator
for lavt nivå
